refactor: 거대한 컴포넌트 분리
