+++ .github/pull_request_template.md
@@ -3,13 +3,14 @@
 ### 기본과제
 
 #### 목표 : 상태관리를 이용한 적절한 분리와 계층에 대한 이해를 통한 FSD 폴더 구조 적용하기
+
 - 관심사를 하나만 가지고 있는가!?
 - 관심사의 분리에 대한 이해
 - FSD(Feature-Sliced Design)에 대한 이해
 
 #### 체크포인트
 
-- [ ] shared 공통 컴포넌트를 분리했나요?
+- [X] shared 공통 컴포넌트를 분리했나요?
 - [ ] shared 공통 로직을 분리했나요?
 - [ ] entities를 중심으로 type을 정의하고 model을 분리했나요?
 - [ ] entities를 중심으로 ui를 분리했나요?
@@ -17,10 +18,9 @@
 - [ ] feature별로 ui/model을 분리했나요?
 - [ ] widget별로 ui를 분리했나요?
 
-
 ### 심화과제
 
-#### 목표: 서버상태관리 도구인 TanstackQuery를 이용하여 비동기코드를 선언적인 함수형 프로그래밍으로 작성하기 
+#### 목표: 서버상태관리 도구인 TanstackQuery를 이용하여 비동기코드를 선언적인 함수형 프로그래밍으로 작성하기
 
 - useState를 줄여라!
 - TanstackQuery의 사용법에 대한 이해
@@ -33,7 +33,6 @@
 - [ ] 쿼리 키가 적절히 설정되었는가?
 - [ ] useState가 아닌 선언적인 함수형 프로그래밍이 적절히 적용되었는가?
 - [ ] 캐싱과 리프레시 전략이 올바르게 구현되었는가?
-
 
 ## 리뷰 받고 싶은 내용
