+++ src/shared/ui/card/index.ts
@@ -1,0 +1,4 @@
+export { Card } from "./Card.tsx"
+export { CardHeader } from "./CardHeader.tsx"
+export { CardTitle } from "./CardTitle.tsx"
+export { CardContent } from "./CardContent.tsx"
