+++ src/shared/ui/dialog/DialogHeader.tsx
@@ -1,0 +1,6 @@
+import * as React from "react"
+
+export const DialogHeader = ({ className, ...props }: React.HTMLAttributes<HTMLDivElement>) => (
+  <div className={`flex flex-col space-y-1.5 text-center sm:text-left ${className}`} {...props} />
+)
+DialogHeader.displayName = "DialogHeader"
