+++ src/shared/ui/dialog/index.ts
@@ -1,0 +1,4 @@
+export { Dialog } from "./Dialog.tsx"
+export { DialogContent } from "./DialogContent.tsx"
+export { DialogHeader } from "./DialogHeader.tsx"
+export { DialogTitle } from "./Dialogtitle.tsx"
