+++ src/shared/ui/select/index.ts
@@ -1,0 +1,6 @@
+export { Select } from "./Select"
+export { SelectContent } from "./SelectContent"
+export { SelectGroup } from "./SelectGroup"
+export { SelectItem } from "./SelectItem"
+export { SelectTrigger } from "./SelectTrigger"
+export { SelectValue } from "./SelectValue"
