+++ src/shared/ui/table/TableBody.tsx
@@ -1,0 +1,9 @@
+import * as React from "react"
+import { forwardRef } from "react"
+
+export const TableBody = forwardRef<HTMLTableSectionElement, React.HTMLAttributes<HTMLTableSectionElement>>(
+  ({ className, ...props }, ref) => (
+    <tbody ref={ref} className={`[&_tr:last-child]:border-0 ${className}`} {...props} />
+  ),
+)
+TableBody.displayName = "TableBody"
