+++ src/shared/ui/table/TableHeader.tsx
@@ -1,0 +1,7 @@
+import * as React from "react"
+import { forwardRef } from "react"
+
+export const TableHeader = forwardRef<HTMLTableSectionElement, React.HTMLAttributes<HTMLTableSectionElement>>(
+  ({ className, ...props }, ref) => <thead ref={ref} className={`[&_tr]:border-b ${className}`} {...props} />,
+)
+TableHeader.displayName = "TableHeader"
