+++ src/widgets/header/ui/Header.tsx
@@ -1,7 +1,7 @@
 import React from "react"
 import { MessageSquare } from "lucide-react"
 
-const Header: React.FC = () => {
+export const Header: React.FC = () => {
   return (
     <header className="bg-blue-600 text-white p-4 shadow-md">
       <div className="container mx-auto flex justify-between items-center">
@@ -32,5 +32,3 @@ const Header: React.FC = () => {
     </header>
   )
 }
-
-export default Header
